requirements.txt
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
wordcloud>=1.8.0
matplotlib>=3.4.0
PyYAML>=5.4.0
